sidekick-server/pyproject.toml
[tool.poetry]
name = "chatgpt-retrieval-plugin"
version = "0.1.0"
description = ""
authors = ["isafulf <isabella@openai.com>"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
openai = "^0.27.2"
python-dotenv = "^0.21.1"
pydantic = "^1.10.5"
tenacity = "^8.2.1"
tiktoken = "^0.2.0"
numpy = "^1.24.2"
docx2txt = "^0.8"
PyPDF2 = "^3.0.1"
python-pptx = "^0.6.21"
python-multipart = "^0.0.6"
arrow = "^1.2.3"
pinecone-client = "^2.1.0"
weaviate-client = "^3.12.0"
pymilvus = "^2.2.2"
qdrant-client = {version = "^1.0.4", python = "<3.12"}
redis = "4.5.1"
markdownify = "^0.11.6"
playwright = "^1.31.1"

[tool.poetry.scripts]
start = "server.main:start"

[tool.poetry.group.dev.dependencies]
httpx = "^0.23.3"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
